Default to first-is-1 behavior for parallel specs
It is more predictable and works better with interactive test runs to
default to having the first worker set TEST_ENV_NUMBER=1  for
parallel specs, as opposed to setting it to an empty string.

This means parallel specs aren't using the default DB, so it leaves it
available for quick runs of individual specs.

This can be disabled with --first-is-1=false.

Note that this is deliberately the opposite of parallel-tests default,
which defaults `--first-is-1` to false.